Implement CategoricalNB
